requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
environs==11.0.0
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
marshmallow==3.21.2
multidict==6.0.5
packaging==24.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
speedtest-cli==2.1.3
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4